requirements.txt
claude-agent-sdk>=0.1.58
websockets>=14.0
aiosqlite>=0.20.0
pydantic>=2.11.0
pydantic-settings>=2.5.0
pyyaml>=6.0
bcrypt>=4.0.0
apscheduler>=3.10.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-docx>=1.0
openpyxl>=3.1
PyPDF2>=3.0
